Finalizar pricing y añadir librerias de firestore

+++ app/src/main/java/com/example/presupuestosdisa/model/Colores.kt
@@ -2,5 +2,5 @@ package com.example.presupuestosdisa.model
 
 data class Colores(
     val nombre: String = "",
-    val precio: Int = 0
+    val precio: Double = 0.0
 )
